style(crosschain): format code
